requirements.txt
notion-client
imdbinfo
python-dotenv